fix: resolve asset paths, ROM byte order, audio resampling and renderer batching
Los assets nunca cargaban desde el ejecutable: cada modulo llevaba su propia
lista de prefijos relativos y ninguna combinacion funcionaba a la vez para
recomp/build y para el cwd del README. Se centraliza en AssetPaths, que busca
exported_assets hacia arriba desde el binario y desde el cwd.

- assets: nuevo asset_paths.hpp; map_Kd se resuelve contra el directorio del
  MTL (semantica OBJ/MTL) en vez de contra el cwd, que era la razon por la que
  no cargaba ninguna textura de los modelos.
- rom: deteccion de orden de bytes z64/v64/n64 por firma y normalizacion a
  big-endian; cabecera parseada campo a campo (antes memcpy a un struct sin
  byte swap imprimia 0x100080 donde la ROM dice 0x80001000).
- audio: conversion via SDL_AudioStream; los tracks son mono a 22050 Hz y se
  encolaban crudos en un dispositivo estereo de 44100, sonando a ~4x.
- render: una SDL_RenderGeometry por malla en vez de una por triangulo, plano
  de suelo de 800 draw calls a 1, base de camara cacheada por frame, buffers
  reutilizados e indices a uint32_t (el uint16_t desbordaba al fusionar).
- fisica: movimiento relativo a camara, salto por flanco de subida (mantener A
  producia bunny-hop), colision lateral de plataformas consciente de la altura
  y zona muerta radial.
- obj: deduplicacion de vertices por par (posicion, UV), triangulacion en
  abanico y parseo de indices sin excepciones.
- build: NOMINMAX a nivel de target; las macros min/max de windows.h rompian
  std::min y std::numeric_limits<T>::max().
- docs: READMEs reescritos para reflejar el estado real del proyecto.

+++ recomp/src/gbi.hpp
@@ -50,8 +50,11 @@ struct Vertex3D {
     float u, v;
 };
 
+// Indices de 32 bits: al fusionar la geometria real de la ROM en una sola malla
+// se superan facilmente los 65535 vertices, y el uint16_t anterior desbordaba en
+// silencio produciendo triangulos que apuntaban a vertices arbitrarios.
 struct Triangle3D {
-    uint16_t v0, v1, v2;
+    uint32_t v0, v1, v2;
 };
 
 } // namespace N64
